+++ lib/src/iterable/random.dart
@@ -1,0 +1,24 @@
+import 'dart:math' show Random;
+
+final _defaultRandom = Random();
+
+extension RandomExtension<E> on Iterable<E> {
+  /// Returns a random element from this [Iterable].
+  ///
+  /// For example
+  ///
+  ///    [1, 2, 3, 4, 5, 6].atRandom()
+  ///
+  /// returns a random dice roll each time you call it.
+  ///
+  E atRandom({Random? random, E Function()? orElse}) {
+    if (isNotEmpty) {
+      final index = (random ?? _defaultRandom).nextInt(length);
+      return elementAt(index);
+    }
+    if (orElse == null) {
+      throw StateError('Unable to get random element from empty collection.');
+    }
+    return orElse();
+  }
+}